logging/pyproject.toml
[tool.poetry]
name = "logging"
version = "0.1.0"
description = ""
authors = ["Alex Wendland <a.p.wendland@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
